requirements.txt
fastapi
uvicorn
xgboost
pandas
numpy
python-dotenv
supabase-py
